Добавил модуль для вспомогательных функций, переместил туда функции для получения массивов и создания объектов с данными авточехлов
