+++ js/main.js
@@ -7,6 +7,8 @@
 ////  +2.4. Добавить события для вывода значений в разметке
 //  2. Сделать итоговую разметку
 
+import {getArray, getArrayTextContent, getObj} from "./util.js";
+
 const selectMaterial = document.querySelector('.select-item__material');
 const selectMaterialOptions = selectMaterial.querySelector('option');
 const selectColor = document.querySelector('.select-item__color');
@@ -15,65 +17,15 @@ const selectColorOptions = selectColor.querySelector('option');
 const materialText = document.querySelector('.add-cover-text__material');
 const colorText = document.querySelector('.add-cover-text__color');
 
-function getMaterials () {
-  const materials = [];
-  
-  const materialOptions = document.querySelectorAll('.select-item__material option');
- 
-  materialOptions.forEach((material) => materials.push(material.value));
-  
-  return materials;
-}
-
-function getColors () {
-  const colors = [];
-  
-  const colorOptions = document.querySelectorAll('.select-item__color option');
- 
-  colorOptions.forEach((color) => colors.push(color.value));
-  
-  return colors;
-}
-
-const materials = getMaterials();
+const materials = getArray('.select-item__material option');
+const colors = getArray('.select-item__color option');
 console.log(materials);
-const colors = getColors();
 console.log(colors);
 
-function getMaterialsTextContent () {
-  const materialsText = [];
-  
-  const selectMaterialOptions = selectMaterial.querySelectorAll('option');
-  
-  selectMaterialOptions.forEach((material) => materialsText.push(material.textContent));
-  
-  return materialsText;
-}
-
-function getColorsTextContent () {
-  const colorsText = [];
-  
-  const selectColorOptions = selectColor.querySelectorAll('option');
-  
-  selectColorOptions.forEach((color) => colorsText.push(color.textContent));
-  
-  return colorsText;
-}
-
-const materialsTextContent = getMaterialsTextContent();
-const colorsTextContent = getColorsTextContent();
+const materialsTextContent = getArrayTextContent('.select-item__material option');
+const colorsTextContent = getArrayTextContent('.select-item__color option');
 console.log(materialsTextContent);
 console.log(colorsTextContent);
-
-function getObj (valuesArr, textArr) {
-  const newObj = {}
-  
-  for (let i = 0; i < valuesArr.length; i++) {
-    newObj[valuesArr[i]] = textArr[i]
-  }
-  
-  return newObj;
-}
 
 const materialsObj = getObj(materials, materialsTextContent);
 const colorsObj = getObj(colors, colorsTextContent);
@@ -98,7 +50,6 @@ if (selectColorOptions) {
 
 function materialChangeHandler (evt) {
   materialText.textContent = newCarCover.materials[evt.target.value];
-  console.log(evt.target.value);
 }
 
 function colorChangeHandler (evt) {
